tools/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.22.0
openai==0.28.0
pinecone-client==2.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
yarl==1.9.2
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
argilla==1.10.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
dill==0.3.6
dnspython==2.4.2
docutils==0.20.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.97.0
filetype==1.2.0
Flask==2.3.3
fonttools==4.39.3
freegames==2.5.3
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.171
libmagic==1.0
loguru==0.7.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
ntlm-auth==1.5.0
numexpr==2.8.4
olefile==0.46
omegaup==2.0.14
openai==0.28.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pbr==5.11.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.4
platformdirs==3.6.0
ply==3.11
protobuf==4.22.5
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.2
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
qiskit==0.43.0
qiskit-aer==0.12.0
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.24.0
regex==2023.5.5
requests==2.31.0
requests-ntlm==1.1.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
rustworkx==0.12.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
soupsieve==2.4.1
spicy==0.16.0
SQLAlchemy==2.0.13
starlette==0.27.0
stevedore==5.0.0
symengine==0.9.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tk==0.1.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.7.6
uvicorn==0.22.0
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
argilla==1.10.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
dill==0.3.6
dnspython==2.4.2
docutils==0.20.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.97.0
filetype==1.2.0
Flask==2.3.3
fonttools==4.39.3
freegames==2.5.3
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-documentai==2.18.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.171
libmagic==1.0
loguru==0.7.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
ntlm-auth==1.5.0
numexpr==2.8.4
olefile==0.46
omegaup==2.0.14
openai==0.28.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pbr==5.11.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.4
platformdirs==3.6.0
ply==3.11
proto-plus==1.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyGithub==1.58.2
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
qiskit==0.43.0
qiskit-aer==0.12.0
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.24.0
regex==2023.5.5
requests==2.31.0
requests-ntlm==1.1.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
rustworkx==0.12.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
soupsieve==2.4.1
spicy==0.16.0
SQLAlchemy==2.0.13
starlette==0.27.0
stevedore==5.0.0
symengine==0.9.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tk==0.1.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.7.6
urllib3==1.26.16
uvicorn==0.22.0
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
